# Optimize the migration cache